docs(lib): adds CASL mongoose
I added a library for authorization in Nodejs with MongoDB. The library is called CASL - https://github.com/stalniy/casl. There are also a lot of interesting stuff in  documentation - https://stalniy.github.io/casl/ and [related articles on Medium](https://medium.com/@sergiy.stotskiy/latest). 

The library has been around for almost a year, I used it on several projects and believe it's a very good addition to this list.